pyproject.toml
[project]
name = "the-wrapping-book"
version = "0.1.0"
description = "Repository to create a book using content from thewrappinggallery.com"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cairosvg>=2.7.1",
    "pillow>=11.2.1",
    "python-decouple>=3.8",
    "qrcode>=8.1",
    "reportlab>=4.3.1",
    "requests>=2.32.3",
    "supabase>=2.15.0",
]
